# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "URBIS API",
    "version": "0.5.0-alpha.1",
    "description": "API da plataforma URBIS. Propriedade: Sigma SIHF Soluções Analíticas S/A — CNPJ 01.851.824/0001-38. A especificação diferencia funcionalidades implementadas de dependências externas que ainda exigem validação em staging."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Backend NestJS local"
    },
    {
      "url": "http://localhost:3100",
      "description": "Prova local em memória"
    }
  ],
...
  "paths": {
...
    "/v1/occurrences": {
      "post": {
        "summary": "Criar ocorrência georreferenciada",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Occurrences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "categoryCode",
                  "latitude",
                  "longitude"
                ],
                "properties": {
                  "categoryCode": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "gpsAccuracyM": {
                    "type": "number"
                  },
                  "addressText": {
                    "type": "string"
                  },
                  "cep": {
                    "type": "string"
                  },
                  "clientRequestId": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "NORMAL",
                      "MEDIUM",
                      "HIGH",
                      "CRITICAL"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar ocorrências operacionais",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Occurrences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "neighborhood",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/occurrences/map": {
      "get": {
        "summary": "Consultar ocorrências para mapa",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Occurrences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "neighborhood",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/occurrences/{id}": {
      "get": {
        "summary": "Detalhar ocorrência, timeline e OS",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Occurrences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/occurrences/{id}/triage": {
      "post": {
        "summary": "Triar ocorrência",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Occurrences"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "decision"
                ],
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "PROCEDENT",
                      "IMPROCEDENT",
                      "DUPLICATE"
                    ]
                  },
                  "duplicateOf": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "responsibleOrgCode": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "NORMAL",
                      "MEDIUM",
                      "HIGH",
                      "CRITICAL"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/occurrences/{occurrenceId}/evidence": {
      "get": {
        "summary": "Listar evidências autorizadas",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Evidence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "occurrenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Enviar evidência com SHA-256",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Evidence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "occurrenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/occurrences/{occurrenceId}/evidence/{mediaId}/authority-validate": {
      "post": {
        "summary": "Validar evidência por autoridade",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Evidence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "occurrenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/occurrences/{occurrenceId}/ici": {
      "get": {
        "summary": "Consultar ICI-URBIS",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Evidence"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "occurrenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v1/work-orders": {
      "get": {
        "summary": "Listar ordens de serviço",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "Criar ordem de serviço",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/work-orders/{id}/assign-team": {
      "post": {
        "summary": "Atribuir equipe à OS",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/work-orders/{id}/status": {
      "post": {
        "summary": "Atualizar status de OS pela Central",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/service-orders": {
      "post": {
        "summary": "Criar ordem de serviço (compatibilidade)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/v1/service-orders/team/{teamId}": {
      "get": {
        "summary": "Listar OS por equipe (compatibilidade)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/service-orders/{id}/assign-team": {
      "post": {
        "summary": "Atribuir equipe (compatibilidade)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Work Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/v1/field/work-orders/{id}/status": {
      "post": {
        "summary": "Atualizar OS no campo com geofence",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Field"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
}